centralize setTermBins into termsetting edit menu option handler; fix getData for numeric term that's missing bins object
